new_dynarec: ARM64 read_nomemb_new

+++ src/r4300/new_dynarec/arm64/assem_arm64.c
@@ -1835,7 +1835,7 @@ static void emit_testimm(int rs,int imm)
   output_w32(0x72000000|armval<<10|rs<<5|WZR);
 }
 
-static void emit_testimm64(int rs,int imm)
+static void emit_testimm64(int rs,int64_t imm)
 {
   assert(rs!=29);
   u_int armval, ret;
